refactor(partners): migrate application domain into modules/ + relocate post-install
