+++ packages/twenty-apps/internal/twenty-partners/src/modules/application/__tests__/on-application-set-name.test.ts
