+++ packages/twenty-apps/internal/twenty-partners/src/modules/application/fields/applications-on-partner.field.ts
@@ -5,7 +5,7 @@ import {
   APPLICATIONS_ON_PARTNER_FIELD_ID,
   APPLICATION_OBJECT_UNIVERSAL_IDENTIFIER,
   APPLICATION_PARTNER_FIELD_ID,
-} from 'src/objects/application.object';
+} from 'src/modules/application/objects/application.object';
 
 export default defineField({
   universalIdentifier: APPLICATIONS_ON_PARTNER_FIELD_ID,
