+++ packages/twenty-apps/internal/twenty-partners/src/modules/application/graphql/mutations/delete-application.ts
@@ -1,0 +1,10 @@
+import { type CoreApiClient } from 'twenty-client-sdk/core';
+
+export function deleteApplication(client: CoreApiClient, id: string) {
+  return client.mutation({
+    deleteApplication: {
+      __args: { id },
+      id: true,
+    },
+  });
+}
