+++ packages/twenty-apps/internal/twenty-partners/src/modules/application/graphql/mutations/update-application.ts
@@ -1,0 +1,11 @@
+import { type CoreApiClient, type CoreSchema } from 'twenty-client-sdk/core';
+
+export function updateApplication(
+  client: CoreApiClient,
+  id: string,
+  data: CoreSchema.ApplicationUpdateInput,
+) {
+  return client.mutation({
+    updateApplication: { __args: { id, data }, id: true },
+  });
+}
