+++ packages/twenty-apps/internal/twenty-partners/src/modules/application/graphql/queries/find-application-with-relations.ts
@@ -1,0 +1,12 @@
+import { type CoreApiClient } from 'twenty-client-sdk/core';
+
+export function findApplicationWithRelations(client: CoreApiClient, id: string) {
+  return client.query({
+    application: {
+      __args: { filter: { id: { eq: id } } },
+      id: true,
+      partner: { name: true },
+      opportunity: { name: true },
+    },
+  });
+}
