+++ packages/twenty-apps/internal/twenty-partners/src/modules/application/graphql/queries/find-duplicate-application.ts
@@ -1,0 +1,20 @@
+import { type CoreApiClient } from 'twenty-client-sdk/core';
+
+export function findDuplicateApplication(
+  client: CoreApiClient,
+  opportunityId: string,
+  partnerId: string,
+) {
+  return client.query({
+    applications: {
+      __args: {
+        filter: {
+          opportunityId: { eq: opportunityId },
+          partnerId: { eq: partnerId },
+        },
+        first: 1,
+      },
+      edges: { node: { id: true } },
+    },
+  });
+}
