+++ packages/twenty-apps/internal/twenty-partners/src/modules/application/graphql/queries/find-partner-by-member.ts
@@ -1,0 +1,10 @@
+import { type CoreApiClient } from 'twenty-client-sdk/core';
+
+export function findPartnerByMember(client: CoreApiClient, memberId: string) {
+  return client.query({
+    partners: {
+      __args: { filter: { partnerUserId: { eq: memberId } }, first: 1 },
+      edges: { node: { id: true } },
+    },
+  });
+}
