+++ packages/twenty-apps/internal/twenty-partners/src/modules/application/navigation-menu-items/followup-applications.navigation-menu-item.ts
@@ -1,6 +1,6 @@
 import { NavigationMenuItemType, defineNavigationMenuItem } from 'twenty-sdk/define';
 
-import { FOLLOWUP_APPLICATIONS_VIEW_UNIVERSAL_IDENTIFIER } from 'src/views/followup-applications.view';
+import { FOLLOWUP_APPLICATIONS_VIEW_UNIVERSAL_IDENTIFIER } from 'src/modules/application/views/followup-applications.view';
 
 export default defineNavigationMenuItem({
   universalIdentifier: '636cddf9-b104-432d-a5d3-3bad36b7a54a',
