+++ packages/twenty-apps/internal/twenty-partners/src/modules/application/views/application-record-page-fields.view.ts
@@ -7,7 +7,7 @@ import {
   APPLICATION_PARTNER_FIELD_ID,
   APPLICATION_PITCH_FIELD_ID,
   APPLICATION_STATE_FIELD_ID,
-} from 'src/objects/application.object';
+} from 'src/modules/application/objects/application.object';
 
 export const APPLICATION_RECORD_PAGE_FIELDS_VIEW_ID =
   'e004c2ff-462a-45d4-8a77-071dcf093879';
