+++ packages/twenty-apps/internal/twenty-partners/src/modules/opportunity/fields/applications-on-opportunity.field.ts
@@ -9,7 +9,7 @@ import {
   APPLICATIONS_ON_OPPORTUNITY_FIELD_ID,
   APPLICATION_OBJECT_UNIVERSAL_IDENTIFIER,
   APPLICATION_OPPORTUNITY_FIELD_ID,
-} from 'src/objects/application.object';
+} from 'src/modules/application/objects/application.object';
 
 export default defineField({
   universalIdentifier: APPLICATIONS_ON_OPPORTUNITY_FIELD_ID,
