+++ packages/twenty-apps/internal/twenty-partners/src/modules/shared/post-install.logic-function.ts
